chore: [344] Reverted changes to checkActionForMisplacedClassDecorators
